Update gefs spatial script file

+++ scripts/global_ens/plots/exevs_global_ens_atmos_gefs_precip_spatial_plots.sh
@@ -54,14 +54,20 @@ for model in $model_list ; do
  for fhr in 024 048 072 096 120 144 168 192 216 240 ; do
   past=`$NDATE -$fhr ${VDATE}12`	
   INITDATE=${past:0:8}
-  ln -sf $source/GenEnsProd_${MODEL}_APCP24_FHR${fhr}_${VDATE}_120000V_ens.nc $target/${model}_precip_24hrAccum_init${INITDATE}12_fhr${fhr}.nc	
+  apcp24mean=$source/GenEnsProd_${MODEL}_APCP24_FHR${fhr}_${VDATE}_120000V_ens.nc
+  size=`ls -l $apcp24mean | awk '{ print $5}'`
+  if [ $size -gt 1000000 ] ; then
+    ln -sf $source/GenEnsProd_${MODEL}_APCP24_FHR${fhr}_${VDATE}_120000V_ens.nc $target/${model}_precip_24hrAccum_init${INITDATE}12_fhr${fhr}.nc	
+  fi 
  done
 done
 
 source=$COMINccpa24/gefs
 target=$DATA/grid2grid_plots/data/ccpa
 mkdir -p $target
-ln -sf $source/ccpa.t12z.grid3.24h.f00.nc $target/ccpa_precip_24hrAccum_valid${VDATE}12.nc
+if [ -s $source/ccpa.t12z.grid3.24h.f00.nc ] ; then
+ ln -sf $source/ccpa.t12z.grid3.24h.f00.nc $target/ccpa_precip_24hrAccum_valid${VDATE}12.nc
+fi 
 
 python $USHevs/global_ens/ush_gens_plot_py/global_det_atmos_plots.py
